Working 'up' command.
`minecloud up cliff-side` will now start and EC2 instance, bootstrap that
instance, download the server from S3 and spin up the Minecraft server wrapper
on it.

From here I need to make it so I can do the reverse operation, safely stopping
the server, saving it to S3, and terminating the instance.