Create basic lambda function, fix SSH usage as a result.
Here I create a basic lambda function that starts up cliff-side. I ran into a
lot of issues while doing this.

(1) getting permissions correct for lambda functions to access the EC2 instances
was difficult. I ended up opening the key-only SSH to the world. It might be
that I can fix this with VPCs. It's probably secure enough for now.

(2) For the CLI we would accept a new SSH in certain situations. But we fudged
it and made a subshell that called ssh-keyscan to add the key to the known hosts
file. This didn't work for lambda since it has no such thing installed. I had to
to it properly in pure Go instead and create a brand new known hosts file for it
to work properly.

(3) I had to rejig how Minecloud does configuration, it was too dependent on
running on a normal machine with typical things installed. It is now a lot more
agnostic, but with a bit more setup.