Allow execution of lambda code locally.
This sets up an `Invoker` interface that has two primary implementations. One of
which calls out to lambda functions stored on AWS, and another runs as close as
possible to the same code locally. This mostly allows real testing of the lambda
code without the change-upload-test cycle. The surface of code that lambda runs
that the local version doesn't is minimised.